Правки после консультации
