+++ src/view/cost.js
@@ -1,4 +1,4 @@
-export function createCostView(cost) {
+export function createCost(cost) {
   return `<p class="trip-info__cost">
     Total: &euro;&nbsp;<span class="trip-info__cost-value">${cost}</span>
     </p>`;
