+++ src/view/info-date.js
@@ -1,0 +1,3 @@
+export function createTripInfoDate(dateRange) {
+    return `<p class="trip-info__dates">${dateRange}</p>`;
+}
