+++ src/view/info-title.js
@@ -1,0 +1,3 @@
+export function createTripInfoTitle(title) {
+    return `<h1 class="trip-info__title">${title}</h1>`;
+}
