+++ src/view/switches.js
