--- src/view/trip-info-date-view.js
@@ -1,3 +1,0 @@
-export function createTripInfoDate(range) {
-    return `<p class="trip-info__dates">${range}</p>`;
-}
