--- src/view/trip-info-title-view.js
@@ -1,5 +1,0 @@
-export function createTripInfoTitle(title) {
-    return `
-    <h1 class="trip-info__title">${title}</h1>
-    `;
-}
