+++ src/view/trip-info.js
@@ -1,13 +1,15 @@
 import dayjs from 'dayjs';
 
-import {createTripInfoTitle} from './trip-info-title-view';
-import {createTripInfoDate} from './trip-info-date-view';
-import {createCostView} from './cost-view';
+import { createTripInfoTitle } from './info-title';
+import { createTripInfoDate } from './info-date';
+import { createCost } from './cost';
 
 export function createTripInfo(tripPoints) {
   const firstPoint = tripPoints[0];
   const lastPoint = tripPoints[tripPoints.length - 1];
-  const tripRange = `${dayjs(firstPoint.dates[0]).format('MMM DD')}&nbsp;&mdash;&nbsp;${dayjs(lastPoint.dates[0]).format('DD')}`;
+  const tripRange = dayjs(lastPoint.dates[0]).diff(dayjs(firstPoint.dates[0]), 'month') > 0
+    ? `${dayjs(firstPoint.dates[0]).format('MMM DD')}&nbsp;&mdash;&nbsp;${dayjs(lastPoint.dates[0]).format('MMM DD')}`
+    : `${dayjs(firstPoint.dates[0]).format('MMM DD')}&nbsp;&mdash;&nbsp;${dayjs(lastPoint.dates[0]).format('DD')}`;
 
   let tripTitle;
   if (tripPoints.length > 3) {
@@ -26,6 +28,6 @@ export function createTripInfo(tripPoints) {
           ${createTripInfoTitle(tripTitle)}
           ${createTripInfoDate(tripRange)}
       </div>
-      ${createCostView(cost)}
+      ${createCost(cost)}
   </section>`;
 }
